# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Billing and Payment Service for KelolaKelas Platform",
        "title": "KelolaKelas Billing Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/billing/transactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Parents see their own transactions. Tenant members need the `billing:read` permission in their tenant: without it the answer is 403, and while identity cannot be asked it is 503.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "List billing transactions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.HTTPResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.TransactionListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/billing/transactions/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Same authorization as the list: parents read their own transaction, tenant members need `billing:read` (403 without it, 503 while identity cannot be asked).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "Get billing transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.HTTPResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.TransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/billing/webhooks/duitku": {
            "post": {
                "description": "Callback endpoint for Duitku payment status updates",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "Handle Duitku payment webhook callback",
                "parameters": [
                    {
                        "description": "Duitku callback payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.DuitkuCallbackPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.HTTPResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/internal/billing/transactions": {
            "post": {
                "description": "Internal service-to-service endpoint for generating a billing invoice.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "Generate subscription payment from an internal service",
                "parameters": [
                    {
                        "description": "Payment request details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.GenerateSubscriptionPaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.HTTPResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.GenerateSubscriptionPaymentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/internal/billing/transactions/cancel": {
            "post": {
                "description": "Internal service-to-service endpoint that marks the unpaid transaction of an enrollment as `cancelled`. The transition is idempotent, never rewrites a paid or refunded transaction, and reports 404 when the enrollment has no transaction at all.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "Cancel the unpaid transaction of a cancelled enrollment",
                "parameters": [
                    {
                        "description": "Enrollment to cancel",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.CancelEnrollmentPaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.HTTPResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.TransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_kelolakelas_kelolakelas-billing-service_internal_domain.CancelEnrollmentPaymentRequest": {
            "type": "object",
            "required": [
                "enrollment_id"
            ],
            "properties": {
                "enrollment_id": {
                    "type": "string"
                }
            }
        },
        "github_com_kelolakelas_kelolakelas-billing-service_internal_domain.DuitkuCallbackPayload": {
            "type": "object",
            "required": [
                "amount",
                "merchantCode",
                "merchantOrderId",
                "resultCode",
                "signature"
            ],
            "properties": {
                "amount": {
                    "type": "string"
                },
                "merchantCode": {
                    "type": "string"
                },
                "merchantOrderId": {
                    "type": "string"
                },
                "merchantUserId": {
                    "type": "string"
                },
                "paymentCode": {
                    "type": "string"
                },
                "productDetail": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                },
                "resultCode": {
                    "type": "string"
                },
                "settlementDate": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                }
            }
        },
        "github_com_kelolakelas_kelolakelas-billing-service_internal_domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_kelolakelas_kelolakelas-billing-service_internal_domain.GenerateSubscriptionPaymentRequest": {
            "type": "object",
            "required": [
                "billing_cycle",
                "class_id",
                "enrollment_id",
                "parent_id",
                "student_id",
                "subtotal_amount",
                "tenant_id"
            ],
            "properties": {
                "billing_cycle": {
                    "type": "string",
                    "enum": [
                        "monthly",
                        "quarterly",
                        "yearly"
                    ]
                },
                "class_id": {
                    "type": "string"
                },
                "discount_amount": {
                    "type": "integer",
                    "minimum": 0
                },
                "enrollment_id": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "payment_gateway_fee": {
                    "type": "integer",
                    "minimum": 0
                },
                "platform_fee": {
                    "type": "integer",
                    "minimum": 0
                },
                "sender_email": {
                    "type": "string"
                },
                "sender_name": {
                    "type": "string"
                },
                "sender_phone": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                },
                "subtotal_amount": {
                    "type": "integer"
                },
                "tenant_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "voucher_id": {
                    "type": "string"
                }
            }
        },
        "github_com_kelolakelas_kelolakelas-billing-service_internal_domain.GenerateSubscriptionPaymentResponse": {
            "type": "object",
            "properties": {
                "checkout_session_url": {
                    "type": "string"
                },
                "gross_amount": {
                    "type": "integer"
                },
                "payment_intent_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                }
            }
        },
        "github_com_kelolakelas_kelolakelas-billing-service_internal_domain.HTTPResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "github_com_kelolakelas_kelolakelas-billing-service_internal_domain.TransactionListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kelolakelas_kelolakelas-billing-service_internal_domain.TransactionResponse"
                    }
                },
                "pagination": {
                    "type": "object",
                    "properties": {
                        "page": {
                            "type": "integer"
                        },
                        "page_size": {
                            "type": "integer"
                        },
                        "total_items": {
                            "type": "integer"
                        },
                        "total_pages": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "github_com_kelolakelas_kelolakelas-billing-service_internal_domain.TransactionResponse": {
            "type": "object",
            "properties": {
                "checkout_session_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "discount_amount": {
                    "type": "integer"
                },
                "enrollment_id": {
                    "type": "string"
                },
                "expired_at": {
                    "type": "string"
                },
                "gross_amount": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "invoice_claimed_at": {
                    "type": "string"
                },
                "invoice_expires_at": {
                    "type": "string"
                },
                "invoice_failure_reason": {
                    "type": "string"
                },
                "merchant_order_id": {
                    "type": "string"
                },
                "net_amount": {
                    "type": "integer"
                },
                "paid_at": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "payment_gateway_fee": {
                    "type": "integer"
                },
                "payment_gateway_provider": {
                    "type": "string"
                },
                "payment_intent_id": {
                    "type": "string"
                },
                "platform_fee": {
                    "type": "integer"
                },
                "reconciliation_attempts": {
                    "type": "integer"
                },
                "reconciliation_kind": {
                    "type": "string"
                },
                "reconciliation_last_error": {
                    "type": "string"
                },
                "reconciliation_next_attempt_at": {
                    "type": "string"
                },
                "reconciliation_status": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                },
                "subtotal_amount": {
                    "type": "integer"
                },
                "tenant_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "voucher_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}